Move print tests to hello_app suite
